+++ test/integration/render/scenarios/regression/repeated-point-line/scenario.js
@@ -3,7 +3,7 @@ const map = new CartoMap({
     background: 'black'
 });
 
-const source = new carto.source.GeoJSON(sources['line-repeated-point']);
+const source = new carto.source.GeoJSON(sources['multi-polygon-line-repeated-point']);
 const viz = new carto.Viz();
 const layer = new carto.Layer('layer', source, viz);