+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add `getCentroid` method to features
+
 ### Fixed
 - Fix `linear` error when evaluating a date property in a feature.